move update/delete project action in separate modules
